Extend existing Storage and Proxy classes ♻️
