+++ server/app.js
@@ -7,7 +7,8 @@ const express = require('express');
 const helmet = require('helmet');
 const origin = require('./shared/origin');
 const path = require('path');
-const storageProxy = require('./shared/storage/proxy');
+const storage = require('./repository/storage');
+const storageProxy = require('./repository/proxy');
 // eslint-disable-next-line require-sort/require-sort
 require('express-async-errors');
 
@@ -40,7 +41,7 @@ app.use(origin());
 app.use(express.static(path.join(__dirname, '../dist/')));
 if (STORAGE_PATH) app.use(express.static(STORAGE_PATH));
 if (storageProxy.isSelfHosted) {
-  const { proxy: middleware } = require('./shared/storage/proxy/mw');
+  const { proxy: middleware } = require('./shared/storage/proxy/mw')(storage, storageProxy);
   app.use(storageProxy.path, middleware);
 }
