+++ server/shared/content-plugins/elementRegistry.js
@@ -7,8 +7,8 @@ const depd = require('depd');
 const elementsList = require('../../../config/shared/core-elements');
 const hooks = require('./elementHooks');
 const pick = require('lodash/pick');
-const storage = require('../storage');
-const storageProxy = require('../storage/proxy');
+const storage = require('../../repository/storage');
+const storageProxy = require('../../repository/proxy');
 const toCase = require('to-case');
 
 const EXTENSIONS_LIST = '../../../extensions/content-elements/index';
