+++ server/shared/storage/helpers.js
@@ -4,7 +4,7 @@ const { elementRegistry } = require('../content-plugins');
 const get = require('lodash/get');
 const config = require('../../../config/server').storage;
 const Promise = require('bluebird');
-const proxy = require('./proxy');
+const proxy = require('../../repository/proxy');
 const set = require('lodash/set');
 const toPairs = require('lodash/toPairs');
 const values = require('lodash/values');
